Add combine reducers and move BooksForm and BooksList to containers folder
