+++ src/containers/BooksList.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import Book from './Book';
+import Book from '../components/Book';
 
 function BooksList() {
-  const books = useSelector((state) => state);
+  const books = useSelector((state) => state.books);
   const dispatch = useDispatch();
+
   const handleRemoveBook = (book) => {
     dispatch({ type: 'REMOVE_BOOK', book });
   };
+
   return (
     <table>
       <thead>
@@ -19,7 +21,7 @@ function BooksList() {
         </tr>
       </thead>
       <tbody>
-        {books.map((book) => (
+        {books && books.map((book) => (
           <Book key={book.id} book={book} handleRemoveBook={() => handleRemoveBook(book)} />
         ))}
       </tbody>
