+++ src/reducers/books.js
@@ -1,4 +1,8 @@
-const booksReducer = (state, action) => {
+const initialState = [{ id: Math.ceil(Math.random() * 1000), title: 'title1', category: 'Learning' },
+  { id: Math.ceil(Math.random() * 1000), title: 'title412421', category: 'Action' },
+  { id: Math.ceil(Math.random() * 1000), title: 'title43214', category: 'Sci-Fi' }];
+
+const books = (state = initialState, action) => {
   switch (action.type) {
     case 'CREATE_BOOK':
       return [...state, action.book];
@@ -9,4 +13,4 @@ const booksReducer = (state, action) => {
   }
 };
 
-export default booksReducer;
+export default books;
