+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import books from './books';
+
+const reducer = combineReducers({
+  books,
+});
+
+export default reducer;
